requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
distlib==0.3.6
Django==4.0.6
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
filelock==3.7.1
greenlet==2.0.0.post0
MarkupSafe==2.1.1
pipenv==2022.5.2
platformdirs==2.5.2
pycparser==2.21
pymongo==3.12.3
pytz==2022.7.1
six==1.16.0
virtualenv==20.17.1
virtualenv-clone==0.5.7